Added custom variable to let user stop reformat
the projectable-constrain-format variable lets users choose whether they
want to constrain the reformat command to only run on files within the
project, it is true by default, if it is nil, then you can reformat any
file even outside the package.

This seems reasonable as you would not want files outside of the projec
to be reformatted to your current projects settings.

Or maybe you do? Who knows, I'm not your mum.