Error if no project is set
If a user tries to call any of the find-file
commands (e.g. projectable-find-file) when a project isn't set, the
result is as though the user is in an empty project (e.g. the user is
given the choice to choose a file when none exist). Now,
projectable--find-file will error if projectable-id is nil.