Refactored the git find file to have new purpose
Now it will check to see whether the file you're running this command
from is already part of a project, if so, it will cowardly restore that
project from the cache and just use that instead.

Otherwise it changes to the git repo it's just created and calls through
as normal.

If no git repo is found and file is not part of a project, it will error.